refactor: streamline icon paths and consolidate icon management
- Replaced references to pathsHugeIcons.ts with paths.ts in IconBase and PlaceholderImageIcon components for improved organization.
- Introduced paths.ts to centralize icon path definitions, enhancing maintainability and clarity.
- Removed pathsPhosphorIcons.ts to eliminate redundancy and streamline the icon management process.
