+++ components/icon/paths.ts
