--- components/icon/pathsPhosphorIcons.ts
@@ -1,337 +1,0 @@
-/* ===== PHOSPHOR ICON PATHS ===== */
-/* ===== SOCIAL MEDIA ICONS ===== */
-export const twitter = [
-  "M6 5H12L26 27H20L6 5Z",
-  "M14.235 17.9412L6 26.9999",
-  "M26 5L17.765 14.0588",
-];
-export const telegram = [
-  "M10 16.8587L21.2825 26.75C21.4125 26.8646 21.57 26.9435 21.7397 26.979C21.9093 27.0144 22.0853 27.0051 22.2502 26.9521C22.4152 26.899 22.5636 26.804 22.6808 26.6763C22.798 26.5486 22.88 26.3926 22.9187 26.2237L28 4.15246C28.005 4.13033 28.0038 4.10726 27.9967 4.08574C27.9895 4.06421 27.9766 4.04504 27.9594 4.03026C27.9422 4.01549 27.9213 4.00568 27.899 4.00189C27.8766 3.99809 27.8536 4.00045 27.8325 4.00871L2.5 13.9225C2.34274 13.983 2.20935 14.0929 2.11986 14.2357C2.03038 14.3785 1.98962 14.5464 2.0037 14.7143C2.01779 14.8823 2.08596 15.0411 2.19799 15.167C2.31002 15.2928 2.45985 15.379 2.625 15.4125L10 16.8587Z",
-  "M10 16.8587L27.9263 4.01123",
-  "M15.5463 21.7225L11.72 25.6925C11.5817 25.836 11.4038 25.9348 11.209 25.9765C11.0141 26.0181 10.8113 26.0006 10.6265 25.9262C10.4417 25.8518 10.2833 25.7239 10.1717 25.5588C10.0601 25.3938 10.0003 25.1992 10 25V16.8588",
-];
-
-export const github = [
-  "M14.9788 7C14.3909 6.0801 13.5809 5.32307 12.6234 4.79876C11.6658 4.27445 10.5917 3.99974 9.5 4C9.0144 4.83956 8.72314 5.77729 8.64765 6.74423C8.57215 7.71117 8.71434 8.68274 9.06375 9.5875C8.37968 10.5949 8.00951 11.7824 8 13V14C8 15.5913 8.63214 17.1174 9.75736 18.2426C10.8826 19.3679 12.4087 20 14 20H20C21.5913 20 23.1174 19.3679 24.2426 18.2426C25.3679 17.1174 26 15.5913 26 14V13C25.9905 11.7824 25.6203 10.5949 24.9363 9.5875C25.2857 8.68274 25.4278 7.71117 25.3524 6.74423C25.2769 5.77729 24.9856 4.83956 24.5 4C23.4083 3.99974 22.3342 4.27445 21.3766 4.79876C20.4191 5.32307 19.6091 6.0801 19.0212 7H14.9788Z",
-  "M13 29V24C13 22.9391 13.4214 21.9217 14.1716 21.1716C14.9217 20.4214 15.9391 20 17 20C18.0609 20 19.0783 20.4214 19.8284 21.1716C20.5786 21.9217 21 22.9391 21 24V29",
-  "M13 26H9C7.93913 26 6.92172 25.5786 6.17157 24.8284C5.42143 24.0783 5 23.0609 5 22C5 20.9391 4.57857 19.9217 3.82843 19.1716C3.07828 18.4214 2.06087 18 1 18",
-];
-
-export const discord = [ // uses fill color
-  "M19.18 9.21123L19.805 6.75748C19.8674 6.5106 20.0206 6.29631 20.2339 6.15731C20.4473 6.01831 20.7052 5.96481 20.9562 6.00748L25.4613 6.74998C25.6558 6.78157 25.8369 6.86941 25.9822 7.00269C26.1275 7.13596 26.2305 7.30883 26.2787 7.49998L29.97 22.0462C30.0264 22.2749 30.0003 22.5161 29.8962 22.7274C29.7922 22.9387 29.6169 23.1065 29.4013 23.2012L21.0263 26.9137C20.8992 26.9714 20.7616 27.0024 20.622 27.0047C20.4825 27.007 20.344 26.9806 20.215 26.9272C20.0861 26.8737 19.9695 26.7944 19.8726 26.694C19.7756 26.5936 19.7002 26.4744 19.6513 26.3437L18.375 22.8825",
-  "M12.82 9.21123L12.195 6.75748C12.1326 6.5106 11.9794 6.29631 11.7661 6.15731C11.5527 6.01831 11.2948 5.96481 11.0438 6.00748L6.53875 6.74998C6.34416 6.78157 6.16307 6.86941 6.01781 7.00269C5.87254 7.13596 5.76945 7.30883 5.72125 7.49998L2.03 22.0437C1.97361 22.2724 1.99974 22.5136 2.10379 22.7249C2.20783 22.9362 2.38312 23.104 2.59875 23.1987L10.9738 26.9112C11.1008 26.9689 11.2384 26.9999 11.378 27.0022C11.5175 27.0045 11.656 26.9781 11.785 26.9247C11.9139 26.8712 12.0305 26.7919 12.1274 26.6915C12.2244 26.5911 12.2998 26.4719 12.3488 26.3412L13.625 22.8825",
-  "M10 9.78878C11.955 9.25528 13.9735 8.98993 16 9.00003C18.0265 8.98993 20.045 9.25528 22 9.78878",
-  "M22 22.2113C20.045 22.7448 18.0265 23.0102 16 23.0001C13.9735 23.0102 11.955 22.7448 10 22.2113",
-  {
-    path:
-      "M11.5 19C12.3284 19 13 18.3284 13 17.5C13 16.6716 12.3284 16 11.5 16C10.6716 16 10 16.6716 10 17.5C10 18.3284 10.6716 19 11.5 19Z",
-    style: "stroke-none fill-stroke",
-  },
-  {
-    path:
-      "M20.5 19C21.3284 19 22 18.3284 22 17.5C22 16.6716 21.3284 16 20.5 16C19.6716 16 19 16.6716 19 17.5C19 18.3284 19.6716 19 20.5 19Z",
-    style: "stroke-none fill-stroke",
-  },
-];
-
-/*
-export const instagram = [ // uses fill color
-  "M16 21C18.7614 21 21 18.7614 21 16C21 13.2386 18.7614 11 16 11C13.2386 11 11 13.2386 11 16C11 18.7614 13.2386 21 16 21Z",
-  "M22 4H10C6.68629 4 4 6.68629 4 10V22C4 25.3137 6.68629 28 10 28H22C25.3137 28 28 25.3137 28 22V10C28 6.68629 25.3137 4 22 4Z",
-  {
-    path:
-      "M22.5 11C23.3284 11 24 10.3284 24 9.5C24 8.67157 23.3284 8 22.5 8C21.6716 8 21 8.67157 21 9.5C21 10.3284 21.6716 11 22.5 11Z",
-    style: "stroke-none fill-stroke",
-  },
-];
-*/
-
-export const globe = [
-  "M16 28C22.6274 28 28 22.6274 28 16C28 9.37258 22.6274 4 16 4C9.37258 4 4 9.37258 4 16C4 22.6274 9.37258 28 16 28Z",
-  "M21 16C21 24 16 28 16 28C16 28 11 24 11 16C11 8 16 4 16 4C16 4 21 8 21 16Z",
-  "M4.68262 12H27.3176",
-  "M4.68262 20H27.3176",
-];
-
-export const envelope = [
-  "M4 7H28V24C28 24.2652 27.8946 24.5196 27.7071 24.7071C27.5196 24.8946 27.2652 25 27 25H5C4.73478 25 4.48043 24.8946 4.29289 24.7071C4.10536 24.5196 4 24.2652 4 24V7Z",
-  "M28 7L16 18L4 7",
-];
-
-/* ===== LIST ICONS ===== */
-/*
-export const list = [
-  "M5 16H27",
-  "M5 8H27",
-  "M5 24H27",
-];
-*/
-
-export const listAsc = [
-  "M6 16H15",
-  "M6 8H23",
-  "M6 24H13",
-  "M18 21L23 26L28 21",
-  "M23 26V14",
-];
-
-export const listDesc = [
-  "M6 16H15",
-  "M6 8H13",
-  "M6 24H23",
-  "M18 11L23 6L28 11",
-  "M23 6V18",
-];
-
-export const listStar = [
-  "M5 8H27",
-  "M5 16H12",
-  "M5 24H14",
-  "M23 21.8175L26.7087 24L25.7 19.9263L29 17.2013L24.6688 16.8675L23 13L21.3312 16.8675L17 17.2013L20.3 19.9263L19.2913 24L23 21.8175Z",
-];
-
-/* ===== ARROWS - CARET - CHEVRON ICONS ===== */
-export const caretUp = [
-  "M6 20L16 10L26 20",
-];
-
-export const caretDown = [
-  "M26 12L16 22L6 12",
-];
-
-export const caretLeft = [
-  "M20 26L10 16L20 6",
-];
-
-export const caretRight = [
-  "M12 6L22 16L12 26",
-];
-
-export const caretDoubleLeft = [
-  "M25 26L15 16L25 6",
-  "M15 26L5 16L15 6",
-];
-
-export const caretDoubleRight = [
-  "M7 6L17 16L7 26",
-  "M17 6L27 16L17 26",
-];
-
-/* ===== IMAGE ICONS ===== */
-export const image = [
-  "M26 5H6C5.44772 5 5 5.44772 5 6V26C5 26.5523 5.44772 27 6 27H26C26.5523 27 27 26.5523 27 26V6C27 5.44772 26.5523 5 26 5Z",
-  "M12 14C13.1046 14 14 13.1046 14 12C14 10.8954 13.1046 10 12 10C10.8954 10 10 10.8954 10 12C10 13.1046 10.8954 14 12 14Z",
-  "M7.08618 27.0001L20.7924 13.2926C20.8853 13.1996 20.9956 13.1258 21.117 13.0755C21.2384 13.0252 21.3685 12.9993 21.4999 12.9993C21.6313 12.9993 21.7615 13.0252 21.8829 13.0755C22.0043 13.1258 22.1146 13.1996 22.2074 13.2926L26.9999 18.0863",
-];
-export const images = [
-  "M26 5H10C9.44772 5 9 5.44772 9 6V22C9 22.5523 9.44772 23 10 23H26C26.5523 23 27 22.5523 27 22V6C27 5.44772 26.5523 5 26 5Z",
-  "M15 13C16.1046 13 17 12.1046 17 11C17 9.89543 16.1046 9 15 9C13.8954 9 13 9.89543 13 11C13 12.1046 13.8954 13 15 13Z",
-  "M23 23V26C23 26.2652 22.8946 26.5196 22.7071 26.7071C22.5196 26.8946 22.2652 27 22 27H6C5.73478 27 5.48043 26.8946 5.29289 26.7071C5.10536 26.5196 5 26.2652 5 26V10C5 9.73478 5.10536 9.48043 5.29289 9.29289C5.48043 9.10536 5.73478 9 6 9H9",
-  "M12.0862 23.0001L21.7924 13.2926C21.8853 13.1996 21.9956 13.1258 22.117 13.0755C22.2384 13.0252 22.3685 12.9993 22.4999 12.9993C22.6313 12.9993 22.7615 13.0252 22.8829 13.0755C23.0043 13.1258 23.1146 13.1996 23.2074 13.2926L26.9999 17.0863",
-];
-export const imageOut = [ // custom icon
-  "M27 13L26.9988 5.00125L19 5",
-  "M17 15L27 5",
-  "M23 17V26C23 26.2652 22.8946 26.5196 22.7071 26.7071C22.5196 26.8946 22.2652 27 22 27H6C5.73478 27 5.48043 26.8946 5.29289 26.7071C5.10536 26.5196 5 26.2652 5 26V10C5 9.73478 5.10536 9.48043 5.29289 9.29289C5.48043 9.10536 5.73478 9 6 9H15",
-  "m 11,17 c 1.1046,0 2.000001,-0.8954 2.000001,-2 0,-1.1046 -0.895401,-2 -2.000001,-2 -1.1046,0 -2,0.8954 -2,2 0,1.1046 0.8954,2 2,2 z",
-  "M 10.092039,26.99715 16.50622,20.4933 c 0.0929,-0.093 0.2032,-0.1668 0.3246,-0.2171 0.1214,-0.0503 0.2515,-0.0762 0.3829,-0.0762 0.1314,0 0.2616,0.0259 0.383,0.0762 0.1214,0.0503 0.2317,0.1241 0.3245,0.2171 l 4.7925,4.7937",
-];
-
-export const imageCode = [ // custom icon
-  "M26 5H6C5.44772 5 5 5.44772 5 6V26C5 26.5523 5.44772 27 6 27H26C26.5523 27 27 26.5523 27 26V6C27 5.44772 26.5523 5 26 5Z",
-  "m 12.327236,13.894027 -2.5217471,2.101455 2.5217471,2.101455",
-  "m 19.652545,13.894027 2.521747,2.101455 -2.521747,2.101455",
-  "m 17.704567,11.312444 -3.409133,9.375112",
-];
-
-export const imageProfile = [
-  "M16 20C18.7614 20 21 17.7614 21 15C21 12.2386 18.7614 10 16 10C13.2386 10 11 12.2386 11 15C11 17.7614 13.2386 20 16 20Z",
-  "M26 5H6C5.44772 5 5 5.44772 5 6V26C5 26.5523 5.44772 27 6 27H26C26.5523 27 27 26.5523 27 26V6C27 5.44772 26.5523 5 26 5Z",
-  "M7.22241 26.9999C7.67351 25.0091 8.78799 23.2307 10.383 21.9567C11.9779 20.6827 13.9586 19.9888 15.9999 19.9888C18.0412 19.9888 20.0219 20.6827 21.6169 21.9567C23.2118 23.2307 24.3263 25.0091 24.7774 26.9999",
-];
-
-export const cornersOut = [
-  "M21 6H26V11",
-  "M11 26H6V21",
-  "M26 21V26H21",
-  "M6 11V6H11",
-];
-
-/* === COIN ICONS === */
-/*
-export const coins = [
-  "M12 15C17.5228 15 22 12.9853 22 10.5C22 8.01472 17.5228 6 12 6C6.47715 6 2 8.01472 2 10.5C2 12.9853 6.47715 15 12 15Z",
-  "M2 10.5V15.5C2 17.985 6.4775 20 12 20C17.5225 20 22 17.985 22 15.5V10.5",
-  "M8 14.625V19.625",
-  "M22 12.09C26.565 12.5075 30 14.3225 30 16.5C30 18.985 25.5225 21 20 21C17.55 21 15.305 20.6038 13.5662 19.945",
-  "M10 19.91V21.5C10 23.985 14.4775 26 20 26C25.5225 26 30 23.985 30 21.5V16.5",
-  "M24 20.625V25.625",
-  "M16 14.625V25.625",
-];
-*/
-
-export const coinsIn = [ // custom icon
-  "M 19,5 19.0012,12.99875 27,13",
-  "M 19,13.000124 27.000132,5",
-  "m 11.428594,19.925735 c 3.550384,0 6.428594,-1.295169 6.428594,-2.892868 0,-1.597685 -2.87821,-2.892867 -6.428594,-2.892867 C 7.878178,14.14 5,15.435182 5,17.032867 c 0,1.597699 2.878178,2.892868 6.428594,2.892868 z",
-  "m 5,17.032867 v 3.214297 c 0,1.597506 2.878403,2.892868 6.428594,2.892868 3.550191,0 6.428594,-1.295362 6.428594,-2.892868 v -3.214297",
-  "m 8.857157,19.684663 v 3.214296",
-  "m 17.857188,18.055014 c 2.934654,0.268394 5.142876,1.435184 5.142876,2.835011 0,1.597503 -2.878403,2.892866 -6.428594,2.892866 -1.575007,0 -3.018225,-0.254701 -4.136029,-0.678216",
-  "m 10.142875,23.082174 v 1.022147 c 0,1.597506 2.878404,2.892867 6.428595,2.892867 3.550191,0 6.428594,-1.295361 6.428594,-2.892867 v -3.214296",
-  "m 19.142907,23.541819 v 3.214297",
-  "m 14.000032,19.684663 v 7.071453",
-];
-
-export const coinsOut = [ // custom icon
-  "M27 13L26.9988 5.00125L19 5",
-  "M 19,13.000124 27.000132,5",
-  "m 11.428594,19.925735 c 3.550384,0 6.428594,-1.295169 6.428594,-2.892868 0,-1.597685 -2.87821,-2.892867 -6.428594,-2.892867 C 7.878178,14.14 5,15.435182 5,17.032867 c 0,1.597699 2.878178,2.892868 6.428594,2.892868 z",
-  "m 5,17.032867 v 3.214297 c 0,1.597506 2.878403,2.892868 6.428594,2.892868 3.550191,0 6.428594,-1.295362 6.428594,-2.892868 v -3.214297",
-  "m 8.857157,19.684663 v 3.214296",
-  "m 17.857188,18.055014 c 2.934654,0.268394 5.142876,1.435184 5.142876,2.835011 0,1.597503 -2.878403,2.892866 -6.428594,2.892866 -1.575007,0 -3.018225,-0.254701 -4.136029,-0.678216",
-  "m 10.142875,23.082174 v 1.022147 c 0,1.597506 2.878404,2.892867 6.428595,2.892867 3.550191,0 6.428594,-1.295361 6.428594,-2.892867 v -3.214296",
-  "m 19.142907,23.541819 v 3.214297",
-  "m 14.000032,19.684663 v 7.071453",
-];
-
-export const coinsHand = [
-  "M25.5 14C27.433 14 29 12.433 29 10.5C29 8.567 27.433 7 25.5 7C23.567 7 22 8.567 22 10.5C22 12.433 23.567 14 25.5 14Z",
-  "M6 26H2C1.73478 26 1.48043 25.8946 1.29289 25.7071C1.10536 25.5196 1 25.2652 1 25V20C1 19.7348 1.10536 19.4804 1.29289 19.2929C1.48043 19.1054 1.73478 19 2 19H6",
-  "M14 20H18L26.375 18.0738C26.6829 17.9894 27.0061 17.9771 27.3195 18.038C27.6329 18.0989 27.9281 18.2312 28.182 18.4247C28.436 18.6181 28.6419 18.8676 28.7838 19.1536C28.9256 19.4396 28.9996 19.7545 29 20.0738C29.0002 20.4592 28.8929 20.8371 28.6902 21.165C28.4875 21.4929 28.1974 21.7578 27.8525 21.93L23 24L15 26H6V19L9.125 15.875C9.40432 15.5967 9.73577 15.3761 10.1004 15.226C10.465 15.0758 10.8557 14.999 11.25 15H17.5C18.163 15 18.7989 15.2634 19.2678 15.7322C19.7366 16.2011 20 16.837 20 17.5C20 18.1631 19.7366 18.7989 19.2678 19.2678C18.7989 19.7366 18.163 20 17.5 20H14Z",
-  "M22 10.6613C21.3207 10.9826 20.556 11.0767 19.8191 10.9297C19.0821 10.7828 18.4121 10.4026 17.908 9.8453C17.4039 9.28804 17.0925 8.58334 17.0199 7.83541C16.9473 7.08749 17.1174 6.33606 17.5049 5.69227C17.8924 5.04847 18.4769 4.5465 19.1718 4.26058C19.8667 3.97466 20.6352 3.91998 21.3636 4.10463C22.092 4.28929 22.7416 4.70346 23.2164 5.2859C23.6912 5.86835 23.9659 6.58813 24 7.33879",
-];
-
-/* === MATH ICONS === */
-export const x = [
-  "M25 7L7 25",
-  "M25 25L7 7",
-];
-
-export const plus = [
-  "M5 16H27",
-  "M16 5V27",
-];
-
-export const percent = [
-  "M25 7L7 25",
-  "M9.5 13C11.433 13 13 11.433 13 9.5C13 7.567 11.433 6 9.5 6C7.567 6 6 7.567 6 9.5C6 11.433 7.567 13 9.5 13Z",
-  "M22.5 26C24.433 26 26 24.433 26 22.5C26 20.567 24.433 19 22.5 19C20.567 19 19 20.567 19 22.5C19 24.433 20.567 26 22.5 26Z",
-];
-
-/* ===== UI ICONS ===== */
-export const copy = [
-  "M9 5H27V23",
-  "M23 9H5V27H23V9Z",
-];
-
-export const share = [
-  "M18.6363 9.1626L11.3638 13.8376",
-  "M11.3638 18.1626L18.6363 22.8376",
-  "M8 20C10.2091 20 12 18.2091 12 16C12 13.7909 10.2091 12 8 12C5.79086 12 4 13.7909 4 16C4 18.2091 5.79086 20 8 20Z",
-  "M22 29C24.2091 29 26 27.2091 26 25C26 22.7909 24.2091 21 22 21C19.7909 21 18 22.7909 18 25C18 27.2091 19.7909 29 22 29Z",
-  "M22 11C24.2091 11 26 9.20914 26 7C26 4.79086 24.2091 3 22 3C19.7909 3 18 4.79086 18 7C18 9.20914 19.7909 11 22 11Z",
-];
-
-export const gear = [
-  "M16 21C18.7614 21 21 18.7614 21 16C21 13.2386 18.7614 11 16 11C13.2386 11 11 13.2386 11 16C11 18.7614 13.2386 21 16 21Z",
-  "M16.256 25.7638C16.0885 25.7638 15.9198 25.7638 15.756 25.7638L11.7498 28C10.1902 27.4754 8.74363 26.6611 7.48604 25.6L7.47104 21.1C7.38229 20.96 7.29854 20.8187 7.22104 20.6737L3.23729 18.405C2.92351 16.8179 2.92351 15.1846 3.23729 13.5975L7.21729 11.335C7.29854 11.1913 7.38229 11.0487 7.46729 10.9087L7.48729 6.40875C8.74374 5.34455 10.1899 4.52731 11.7498 4L15.7498 6.23625C15.9173 6.23625 16.086 6.23625 16.2498 6.23625L20.2498 4C21.8094 4.52461 23.256 5.33886 24.5135 6.4L24.5285 10.9C24.6173 11.04 24.701 11.1813 24.7785 11.3263L28.7598 13.5938C29.0736 15.1809 29.0736 16.8141 28.7598 18.4013L24.7798 20.6637C24.6985 20.8075 24.6148 20.95 24.5298 21.09L24.5098 25.59C23.2542 26.6544 21.8089 27.472 20.2498 28L16.256 25.7638Z",
-];
-
-export const funnel = [
-  "M4.26259 7.6725C4.13247 7.5293 4.04668 7.35142 4.01564 7.16044C3.9846 6.96946 4.00963 6.77357 4.08769 6.59652C4.16576 6.41948 4.29352 6.26889 4.45546 6.16301C4.61741 6.05713 4.8066 6.0005 5.00009 6H27.0001C27.1938 6.00002 27.3833 6.05629 27.5456 6.16199C27.708 6.26768 27.8361 6.41825 27.9145 6.59538C27.9929 6.77252 28.0181 6.96861 27.9871 7.15982C27.9562 7.35104 27.8704 7.52914 27.7401 7.6725L19.0001 17V24.3312C19.0002 24.4959 18.9596 24.6581 18.882 24.8033C18.8044 24.9485 18.6921 25.0724 18.5551 25.1637L14.5551 27.83C14.4046 27.9304 14.2298 27.9881 14.0491 27.9969C13.8685 28.0058 13.6888 27.9655 13.5292 27.8803C13.3696 27.7952 13.2362 27.6683 13.143 27.5133C13.0498 27.3583 13.0004 27.1809 13.0001 27V17L4.26259 7.6725Z",
-];
-
-export const magnifingGlass = [
-  "M14 24C19.5228 24 24 19.5228 24 14C24 8.47715 19.5228 4 14 4C8.47715 4 4 8.47715 4 14C4 19.5228 8.47715 24 14 24Z",
-  "M21.0713 21.0713L28 28",
-];
-
-export const stamp = [
-  "M5 28H27",
-  "M14.29 17L12.0675 6.625C11.9744 6.18793 11.9802 5.73559 12.0843 5.30101C12.1884 4.86644 12.3883 4.46062 12.6693 4.11319C12.9504 3.76577 13.3055 3.48552 13.7088 3.29292C14.112 3.10033 14.5532 3.00025 15 3H17C17.447 3.00006 17.8883 3.1 18.2917 3.29252C18.6951 3.48504 19.0504 3.76526 19.3316 4.11271C19.6128 4.46015 19.8128 4.86604 19.917 5.30071C20.0212 5.73538 20.0269 6.18783 19.9338 6.625L17.7113 17",
-  "M26 17H6C5.44772 17 5 17.4477 5 18V23C5 23.5523 5.44772 24 6 24H26C26.5523 24 27 23.5523 27 23V18C27 17.4477 26.5523 17 26 17Z",
-];
-
-export const flame = [
-  "M17 24C19.5 23.5788 21.5762 21.5 22 19",
-  "M14 12L17.2838 3C19.9825 5.24 26 11.0188 26 18C26 20.6522 24.9464 23.1957 23.0711 25.0711C21.1957 26.9464 18.6522 28 16 28C13.3478 28 10.8043 26.9464 8.92893 25.0711C7.05357 23.1957 6 20.6522 6 18C6 14.1788 7.8025 10.7175 9.875 8L14 12Z",
-];
-
-export const atom = [ // uses fill color
-  "M26.2819 26.2823C28.4361 24.128 25.579 17.7782 19.9003 12.0995C14.2216 6.42076 7.8717 3.5636 5.71747 5.71783C3.56324 7.87206 6.4204 14.2219 12.0991 19.9006C17.7778 25.5793 24.1277 28.4365 26.2819 26.2823Z",
-  "M19.9003 19.9005C25.579 14.2218 28.4361 7.87197 26.2819 5.71774C24.1277 3.56351 17.7778 6.42066 12.0991 12.0994C6.4204 17.7781 3.56324 24.1279 5.71747 26.2822C7.8717 28.4364 14.2216 25.5792 19.9003 19.9005Z",
-  {
-    path:
-      "M16 17.5C16.8284 17.5 17.5 16.8284 17.5 16C17.5 15.1716 16.8284 14.5 16 14.5C15.1716 14.5 14.5 15.1716 14.5 16C14.5 16.8284 15.1716 17.5 16 17.5Z",
-    style: "stroke-none fill-stroke",
-  },
-];
-
-/* === MEDIA ICONS === */
-// Both icons use fill values and must be defined inline in className - fill-stamp-grey hover:fill-stamp-purple-bright
-export const play = [
-  "M30 16C30.0008 16.3395 29.9138 16.6735 29.7473 16.9694C29.5808 17.2654 29.3406 17.5132 29.05 17.6888L11.04 28.7063C10.7364 28.8922 10.3886 28.9937 10.0326 29.0003C9.67661 29.0069 9.32532 28.9183 9.015 28.7438C8.70764 28.5719 8.4516 28.3213 8.2732 28.0177C8.09481 27.7141 8.00051 27.3684 8 27.0163V4.98376C8.00051 4.63162 8.09481 4.28597 8.2732 3.98235C8.4516 3.67874 8.70764 3.42812 9.015 3.25626C9.32532 3.0817 9.67661 2.99314 10.0326 2.99973C10.3886 3.00632 10.7364 3.10783 11.04 3.29376L29.05 14.3113C29.3406 14.4869 29.5808 14.7347 29.7473 15.0306C29.9138 15.3265 30.0008 15.6605 30 16Z",
-];
-
-export const pause = [
-  "M27 6V26C27 26.5304 26.7893 27.0391 26.4142 27.4142C26.0391 27.7893 25.5304 28 25 28H20C19.4696 28 18.9609 27.7893 18.5858 27.4142C18.2107 27.0391 18 26.5304 18 26V6C18 5.46957 18.2107 4.96086 18.5858 4.58579C18.9609 4.21071 19.4696 4 20 4H25C25.5304 4 26.0391 4.21071 26.4142 4.58579C26.7893 4.96086 27 5.46957 27 6ZM12 4H7C6.46957 4 5.96086 4.21071 5.58579 4.58579C5.21071 4.96086 5 5.46957 5 6V26C5 26.5304 5.21071 27.0391 5.58579 27.4142C5.96086 27.7893 6.46957 28 7 28H12C12.5304 28 13.0391 27.7893 13.4142 27.4142C13.7893 27.0391 14 26.5304 14 26V6C14 5.46957 13.7893 4.96086 13.4142 4.58579C13.0391 4.21071 12.5304 4 12 4Z",
-];
-
-/* ===== SECURITY ICONS ===== */
-export const lockOpen = [
-  "M26 11H6C5.44772 11 5 11.4477 5 12V26C5 26.5523 5.44772 27 6 27H26C26.5523 27 27 26.5523 27 26V12C27 11.4477 26.5523 11 26 11Z",
-  "M11 11V7C11 5.67392 11.5268 4.40215 12.4645 3.46447C13.4021 2.52678 14.6739 2 16 2C18.4188 2 20.5362 3.7175 21 6",
-];
-
-export const lockClosed = [
-  "M26 11H6C5.44772 11 5 11.4477 5 12V26C5 26.5523 5.44772 27 6 27H26C26.5523 27 27 26.5523 27 26V12C27 11.4477 26.5523 11 26 11Z",
-  "M11 11V7C11 5.67392 11.5268 4.40215 12.4645 3.46447C13.4021 2.52678 14.6739 2 16 2C17.3261 2 18.5979 2.52678 19.5355 3.46447C20.4732 4.40215 21 5.67392 21 7V11",
-];
-
-/* === MISC ICONS === */
-export const eye = [
-  "M16 7C6 7 2 16 2 16C2 16 6 25 16 25C26 25 30 16 30 16C30 16 26 7 16 7Z",
-  "M16 21C18.7614 21 21 18.7614 21 16C21 13.2386 18.7614 11 16 11C13.2386 11 11 13.2386 11 16C11 18.7614 13.2386 21 16 21Z",
-];
-
-export const eyeSlash = [
-  "M6 5L26 27",
-  "M19.3637 19.7C18.3824 20.5921 17.0869 21.0579 15.7622 20.9948C14.4375 20.9318 13.1921 20.345 12.3 19.3637C11.4079 18.3824 10.9421 17.0869 11.0052 15.7622C11.0682 14.4375 11.6549 13.1921 12.6362 12.3",
-  "M16.9413 11.0887C18.0043 11.2923 18.9725 11.8353 19.7006 12.6361C20.4286 13.437 20.8772 14.4525 20.9788 15.53",
-  "M26.0763 21.1375C28.8013 18.6975 30 16 30 16C30 16 26 7.00001 16 7.00001C15.1341 6.99883 14.2695 7.06907 13.415 7.21001",
-  "M9.25 8.57501C4.15375 11.155 2 16 2 16C2 16 6 25 16 25C18.3431 25.0184 20.657 24.4785 22.75 23.425",
-];
-
-export const clockCounterClockwise = [
-  "M16 10V16L21 19",
-  "M9 13H4V8",
-  "M8.45 24C10.0225 25.4838 11.9973 26.4713 14.1278 26.8391C16.2583 27.207 18.4499 26.9388 20.4289 26.0682C22.4078 25.1976 24.0864 23.7631 25.2547 21.944C26.4231 20.1249 27.0295 18.0018 26.9982 15.8401C26.9668 13.6783 26.2991 11.5737 25.0785 9.7892C23.8579 8.00472 22.1385 6.6195 20.1351 5.80662C18.1318 4.99375 15.9333 4.78926 13.8144 5.21871C11.6954 5.64815 9.75004 6.6925 8.22125 8.22124C6.75 9.71124 5.535 11.1162 4 13",
-];
-
-/*
-export const pageOut = [
-  "M27 13L26.9988 5.00125L19 5",
-  "M17 15L27 5",
-  "M23 17V26C23 26.2652 22.8946 26.5196 22.7071 26.7071C22.5196 26.8946 22.2652 27 22 27H6C5.73478 27 5.48043 26.8946 5.29289 26.7071C5.10536 26.5196 5 26.2652 5 26V10C5 9.73478 5.10536 9.48043 5.29289 9.29289C5.48043 9.10536 5.73478 9 6 9H15",
-];
-
-export const pageIn = [
-  "M15 25L14.9987 17.0012L7 17",
-  "M5 27L15 17",
-  "M20 23H26C26.2652 23 26.5196 22.8946 26.7071 22.7071C26.8946 22.5196 27 22.2652 27 22V6C27 5.73478 26.8946 5.48043 26.7071 5.29289C26.5196 5.10536 26.2652 5 26 5H10C9.73478 5 9.48043 5.10536 9.29289 5.29289C9.10536 5.48043 9 5.73478 9 6V12",
-];
-
-export const cornersOut = [
-  "M21 6H26V11",
-  "M11 26H6V21",
-  "M26 21V26H21",
-  "M6 11V6H11",
-];
-
-export const cornersIn = [
-  "M26 12H20V6",
-  "M6 20H12V26",
-  "M20 26V20H26",
-  "M12 6V12H6",
-];
-*/
